Adding dotfiles and isaacs file linker
Thanks to Isaac for the file linker via systemd tmp files as a way to
manage your home directory docs (+ among other things).

Going to include my dotfiles repo here as well since this is were the
majority of the inputs will be sourced. This will change the manual
linking was happening previously with stow.