init cntr on driver
cntr is a nice package for helping troublehshot nix builds:

  https://github.com/Mic92/cntr

It required modifications for the root zpool for setting posixacls. This
was done on the entire pool but I would like to better understand the
actual requirements going forward to see if I can enable it on a subset
of the datasets. Checkout the following issue for more information about
why posixacls are required for cntr:

  https://github.com/Mic92/cntr/issues/108